backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
rapidocr-onnxruntime==1.4.4
numpy==2.1.1
pillow==10.4.0
python-multipart==0.0.9
